API/requirements.txt
pandas==2.1.3
scikit_surprise==1.1.3
streamlit==1.28.2
